Fix main

+++ internal/api/record.go
@@ -1,13 +1,13 @@
 package api
 
-// Record represents a record in a specific Zone
+// Record represents a record in a specific Zone.
 type Record struct {
 	ZoneID string `json:"zone_id"`
 	ID     string `json:"id"`
 	Type   string `json:"type"`
 	Name   string `json:"name"`
 	Value  string `json:"value"`
-	TTL    *int   `json:"ttl,omitempty"`
+	TTL    *int64 `json:"ttl,omitempty"`
 }
 
 // HasTTL returns true if a Record has a TTL set and false if TTL is undefined
@@ -15,21 +15,21 @@ func (r *Record) HasTTL() bool {
 	return r.TTL != nil
 }
 
-// CreateRecordRequest represents all data required to create a new record
+// CreateRecordRequest represents all data required to create a new record.
 type CreateRecordRequest struct {
 	ZoneID string `json:"zone_id"`
 	Type   string `json:"type"`
 	Name   string `json:"name"`
 	Value  string `json:"value"`
-	TTL    *int   `json:"ttl,omitempty"`
+	TTL    *int64 `json:"ttl,omitempty"`
 }
 
-// RecordsResponse represents a response from tha API containing a list of records
+// RecordsResponse represents a response from the API containing a list of records.
 type RecordsResponse struct {
 	Records []Record `json:"records"`
 }
 
-// RecordResponse represents a response from the API containing only one record
+// RecordResponse represents a response from the API containing only one record.
 type RecordResponse struct {
 	Record Record `json:"record"`
 }
